broke out waypoint logic into it's own module added handling for mission file when waypoint file is expected and vice versa
